app/requirements.txt
yfinance
schedule
requests
pyyaml

